FIX Allow delay for job health checks
An interim solution for issue #46

Job queue health check can now be configured to extend the duration before the
health check is run. Configure

QueuedJobService:
  health_check_mins: {int}

for the number of minutes to skip on checking. The check is done by modding
this value against the current time's minute, and if that returns 0, then the
health check is performed.